check #1

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,8 @@ use App\Like;
 use App\Post;
 use App\Tag;
 use Illuminate\Http\Request;
+use Auth;
+use Gate;
 
 class PostController extends Controller
 {
@@ -54,11 +56,12 @@ class PostController extends Controller
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        $user = Auth::user();
         $post = new Post([
             'title' => $request->input('title'),
             'content' => $request->input('content')
         ]);
-        $post->save();
+        $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
 
         return redirect()->route('admin.index')->with('info', 'Post created, Title is: ' . $request->input('title'));
@@ -66,11 +69,18 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        
         $post = Post::find($request->input('id'));
+        if (Gate::denies('update-post', $post))
+        {
+            return redirect()->back();
+        }
+        
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -83,6 +93,11 @@ class PostController extends Controller
     public function getAdminDelete($id)
     {
         $post = Post::find($id);
+        if (Gate::denies('update-post', $post)) {
+
+            return redirect()->back();
+
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
